+++ packages/starter-core/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @asteby/metacore-starter-core
 
+## 117.0.0
+
+### Patch Changes
+
+- Updated dependencies [086e2b3]
+  - @asteby/metacore-runtime-react@38.1.0
+
+## 116.0.0
+
+### Patch Changes
+
+- Updated dependencies [5f872f0]
+  - @asteby/metacore-ui@2.18.0
+  - @asteby/metacore-runtime-react@38.0.0
+  - @asteby/metacore-auth@25.0.0
+
+## 115.0.0
+
+### Patch Changes
+
+- Updated dependencies [8231b21]
+  - @asteby/metacore-runtime-react@37.6.0
+
 ## 114.0.0
 
 ### Patch Changes